requirements.txt
gradio>=4.0.0
pandas>=2.0.0
python-dateutil>=2.8.0
numpy>=1.24.0
plotly>=5.0.0
openpyxl>=3.0.0
streamlit>=1.28.0
matplotlib>=3.7.0
mysql-connector-python>=8.1.0
webdriver-manager>=4.0.0